分開acionts的component
